web-app/requirements.txt
fastapi==0.123.8
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
sqlmodel==0.0.27
Jinja2==3.1.6
msal==1.34.0
python-jose[cryptography]==3.5.0
requests==2.32.5
psycopg2-binary==2.9.11

# Document Processing
pdf2image==1.17.0
pdfplumber==0.11.4
python-docx==1.2.0
python-pptx==1.0.2
openpyxl==3.1.5
pillow==12.0.0
python-multipart==0.0.21

# Vector Database & Embeddings
pgvector==0.4.2

# Azure OpenAI
openai==2.14.0

# Azure Blob Storage
azure-storage-blob==12.27.1

# Azure Queue Storage
azure-storage-queue==12.15.0

# Redis for WebSocket pub/sub
redis==7.1.0

# Database Migrations
alembic==1.18.0